[mfp] Fix bug missing tx in epoch executed (#22733)
Previously we did not properly mark the fastpath committed transaction
as executed in epoch. This PR fixes it.
Remove as much special handling of mysticeti-fp transactions as possible
in the execution path.
Only skip process_certificates, which does object loading and execution
in the case.

Updated test to check.

---

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] gRPC:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK: