[move] Fix snapshot of global value updates in traces (#22777)
## Description 

Fixes a bug in the Move trace generation where we were not properly
recording mutations to global values.

This fixes this by re-rooting global values as they move around the
program state. This "re-rooting" tells us where to look for snapshots of
the value of that global at a given time these are:
1. `InLocal(Frame, Local)` -- when generating a snapshot for this value
look into the `Local` in `Frame` for the value.
2. `Value(TraceValue)` this is the value itself. This is when it is on
the stack, or just after being passed in/returned from a native.
3. `OnStack(StackIdx)` is an ephemeral state that is used when local
containing a global reference is `MoveLoc`d to the stack. We need this
for various bookeeping reasons...

These are all runtime state changes though -- the produced trace should
be identical to before except that the snapshots of the global value
will change as the program mutates the value.

## Test plan 

Ran some local tests. Need to work on adding better testing for this,
but may do this in a separate PR.

I think this should be good, but I want to run it on a bunch more
transactions before calling it "good" but it's probably good enough to
start working against in the meantime.